sdk/python/requirements-test.txt
google-api-core==1.*
google-auth==1.*
google-cloud-bigquery==1.*
google-cloud-bigquery-storage==0.*
google-cloud-storage==1.*
google-resumable-media==0.3.1
googleapis-common-protos==1.*
grpcio==1.*
numpy
mock==2.0.0
pandas==0.*
protobuf==3.*
pytest
pytest-mock
pytest-timeout
PyYAML==5.1.2
fastavro==0.21.*
grpcio-testing==1.*
pytest-ordering==0.6.*
pyarrow